rework filesystem & make bios nodes unbreakable
computer data is stored in a directory for the player's computers, giving computers more of ownership and lessening possible name conflicts.

when bios nodes were broken. sometimes a on node would still be placed. when the on node was broken the entire digicompute directory would be removed since the computer-specific directory had already been removed.